+++ tools/archsig/src/lib.rs
@@ -1,15 +1,25 @@
-use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
+use std::collections::{BTreeMap, BTreeSet};
 use std::error::Error;
-use std::ffi::OsStr;
 use std::fs;
-use std::path::{Component as PathComponent, Path, PathBuf};
+use std::path::Path;
 
-use serde::Deserialize;
 use serde_json::Value;
-use walkdir::{DirEntry, WalkDir};
 
+mod extractor;
+mod graph;
+mod policy;
+mod runtime;
 pub mod schema;
+
+pub use extractor::{extract_sig0, extract_sig0_with_policy, extract_sig0_with_runtime};
 pub use schema::*;
+
+use graph::Graph;
+
+#[cfg(test)]
+pub(crate) use extractor::{ParsedImport, parse_imports};
+#[cfg(test)]
+pub(crate) use graph::compute_signature;
 
 const DATASET_SIGNATURE_AXES: [&str; 16] = [
     "hasCycle",
@@ -47,187 +57,6 @@ const DATASET_DELTA_AXES: [&str; 15] = [
     "runtimePropagation",
     "relationComplexity",
 ];
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ImportEdge {
-    source: String,
-    target: String,
-    evidence: String,
-}
-
-pub fn extract_sig0(root: &Path) -> Result<Sig0Document, Box<dyn Error>> {
-    extract_sig0_with_policy(root, None)
-}
-
-pub fn extract_sig0_with_policy(
-    root: &Path,
-    policy_path: Option<&Path>,
-) -> Result<Sig0Document, Box<dyn Error>> {
-    extract_sig0_with_runtime(root, policy_path, None)
-}
-
-pub fn extract_sig0_with_runtime(
-    root: &Path,
-    policy_path: Option<&Path>,
-    runtime_edges_path: Option<&Path>,
-) -> Result<Sig0Document, Box<dyn Error>> {
-    let root = root.to_path_buf();
-    let mut components = Vec::new();
-    let mut import_edges = Vec::new();
-
-    for path in lean_files(&root)? {
-        let relative_path = normalize_relative_path(&root, &path)?;
-        let module_id = module_id_from_relative_path(&relative_path)?;
-        components.push(Component {
-            id: module_id.clone(),
-            path: relative_path,
-        });
-
-        let source = fs::read_to_string(&path)?;
-        for parsed_import in parse_imports(&source) {
-            import_edges.push(ImportEdge {
-                source: module_id.clone(),
-                target: parsed_import.module,
-                evidence: parsed_import.evidence,
-            });
-        }
-    }
-
-    components.sort_by(|a, b| a.id.cmp(&b.id).then(a.path.cmp(&b.path)));
-
-    let edges = dedup_edges(import_edges);
-    let mut signature = compute_signature(&components, &edges);
-
-    let mut metric_status = BTreeMap::new();
-    for axis in ["hasCycle", "sccMaxSize", "maxDepth", "fanoutRisk"] {
-        metric_status.insert(axis.to_string(), measured_status("archsig:import-graph"));
-    }
-    metric_status.insert(
-        "boundaryViolationCount".to_string(),
-        MetricStatus {
-            measured: false,
-            reason: Some("policy file not provided".to_string()),
-            source: None,
-        },
-    );
-    metric_status.insert(
-        "abstractionViolationCount".to_string(),
-        MetricStatus {
-            measured: false,
-            reason: Some("policy file not provided".to_string()),
-            source: None,
-        },
-    );
-    let mut policies = Policies {
-        boundary_allowed: Vec::new(),
-        abstraction_allowed: Vec::new(),
-        policy_id: None,
-        schema_version: None,
-        boundary_group_count: None,
-        abstraction_relation_count: None,
-    };
-    let mut policy_violations = Vec::new();
-
-    if let Some(policy_path) = policy_path {
-        let policy = PolicyFile::read(policy_path)?;
-        policies = Policies {
-            boundary_allowed: Vec::new(),
-            abstraction_allowed: Vec::new(),
-            policy_id: Some(policy.policy_id.clone()),
-            schema_version: Some(policy.schema_version.clone()),
-            boundary_group_count: policy
-                .boundary
-                .as_ref()
-                .map(|boundary| boundary.groups.len()),
-            abstraction_relation_count: policy
-                .abstraction
-                .as_ref()
-                .map(|abstraction| abstraction.relations.len()),
-        };
-        let policy_source = format!("policy:{}", policy.policy_id);
-
-        match policy.boundary.as_ref() {
-            Some(boundary) => match measure_boundary(boundary, &components, &edges) {
-                Ok((count, violations)) => {
-                    signature.boundary_violation_count = count;
-                    metric_status.insert(
-                        "boundaryViolationCount".to_string(),
-                        measured_status(&policy_source),
-                    );
-                    policy_violations.extend(violations);
-                }
-                Err(reason) => {
-                    signature.boundary_violation_count = 0;
-                    metric_status.insert(
-                        "boundaryViolationCount".to_string(),
-                        unmeasured_status(reason),
-                    );
-                }
-            },
-            None => {
-                signature.boundary_violation_count = 0;
-                metric_status.insert(
-                    "boundaryViolationCount".to_string(),
-                    unmeasured_status("boundary policy not provided".to_string()),
-                );
-            }
-        }
-
-        match policy.abstraction.as_ref() {
-            Some(abstraction) => match measure_abstraction(abstraction, &components, &edges) {
-                Ok((count, violations)) => {
-                    signature.abstraction_violation_count = count;
-                    metric_status.insert(
-                        "abstractionViolationCount".to_string(),
-                        measured_status(&policy_source),
-                    );
-                    policy_violations.extend(violations);
-                }
-                Err(reason) => {
-                    signature.abstraction_violation_count = 0;
-                    metric_status.insert(
-                        "abstractionViolationCount".to_string(),
-                        unmeasured_status(reason),
-                    );
-                }
-            },
-            None => {
-                signature.abstraction_violation_count = 0;
-                metric_status.insert(
-                    "abstractionViolationCount".to_string(),
-                    unmeasured_status("abstraction policy not provided".to_string()),
-                );
-            }
-        }
-    }
-
-    let (runtime_edge_evidence, runtime_dependency_graph) =
-        if let Some(runtime_edges_path) = runtime_edges_path {
-            let evidence = RuntimeEdgeEvidenceFile::read(runtime_edges_path, &components)?;
-            let projection = project_runtime_dependency_graph(&evidence);
-            metric_status.insert(
-                "runtimePropagation".to_string(),
-                measured_status(RUNTIME_PROJECTION_RULE_VERSION),
-            );
-            (evidence, Some(projection))
-        } else {
-            (Vec::new(), None)
-        };
-
-    Ok(Sig0Document {
-        schema_version: SCHEMA_VERSION.to_string(),
-        root: root.display().to_string(),
-        component_kind: COMPONENT_KIND.to_string(),
-        components,
-        edges,
-        policies,
-        signature,
-        metric_status,
-        policy_violations,
-        runtime_edge_evidence,
-        runtime_dependency_graph,
-    })
-}
 
 pub fn validate_component_universe_report(
     document: &Sig0Document,
@@ -4167,754 +3996,6 @@ fn unmeasured_status(reason: String) -> MetricStatus {
         measured: false,
         reason: Some(reason),
         source: None,
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct PolicyFile {
-    schema_version: String,
-    policy_id: String,
-    component_id_kind: String,
-    selector_semantics: String,
-    boundary: Option<BoundaryPolicy>,
-    abstraction: Option<AbstractionPolicy>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct RuntimeEdgeEvidenceFile {
-    schema_version: String,
-    component_id_kind: String,
-    edges: Vec<RuntimeEdgeEvidence>,
-}
-
-impl RuntimeEdgeEvidenceFile {
-    fn read(
-        path: &Path,
-        components: &[Component],
-    ) -> Result<Vec<RuntimeEdgeEvidence>, Box<dyn Error>> {
-        let source = fs::read_to_string(path)?;
-        let file: RuntimeEdgeEvidenceFile = serde_json::from_str(&source)?;
-        if file.schema_version != RUNTIME_EDGE_EVIDENCE_SCHEMA_VERSION {
-            return Err(format!(
-                "unsupported runtime edge evidence schemaVersion: {}",
-                file.schema_version
-            )
-            .into());
-        }
-        if file.component_id_kind != COMPONENT_KIND {
-            return Err(format!(
-                "unsupported runtime edge evidence componentIdKind: {}",
-                file.component_id_kind
-            )
-            .into());
-        }
-
-        let component_ids = local_component_ids(components);
-        let mut edges = file.edges;
-        for edge in &edges {
-            validate_runtime_edge_evidence(edge, &component_ids)?;
-        }
-        edges.sort_by(|a, b| {
-            a.source
-                .cmp(&b.source)
-                .then(a.target.cmp(&b.target))
-                .then(a.label.cmp(&b.label))
-                .then(a.evidence_location.path.cmp(&b.evidence_location.path))
-                .then(a.evidence_location.line.cmp(&b.evidence_location.line))
-                .then(a.evidence_location.symbol.cmp(&b.evidence_location.symbol))
-        });
-        Ok(edges)
-    }
-}
-
-fn validate_runtime_edge_evidence(
-    edge: &RuntimeEdgeEvidence,
-    component_ids: &BTreeSet<String>,
-) -> Result<(), String> {
-    if edge.source.is_empty() {
-        return Err("runtime edge source must not be empty".to_string());
-    }
-    if edge.target.is_empty() {
-        return Err("runtime edge target must not be empty".to_string());
-    }
-    if edge.label.is_empty() {
-        return Err("runtime edge label must not be empty".to_string());
-    }
-    if edge.evidence_location.path.is_empty() {
-        return Err("runtime edge evidenceLocation.path must not be empty".to_string());
-    }
-    if !component_ids.contains(&edge.source) {
-        return Err(format!(
-            "runtime edge source is outside component universe: {}",
-            edge.source
-        ));
-    }
-    if !component_ids.contains(&edge.target) {
-        return Err(format!(
-            "runtime edge target is outside component universe: {}",
-            edge.target
-        ));
-    }
-    Ok(())
-}
-
-fn project_runtime_dependency_graph(
-    evidence: &[RuntimeEdgeEvidence],
-) -> RuntimeDependencyGraphProjection {
-    let mut by_pair: BTreeMap<(String, String), usize> = BTreeMap::new();
-    for item in evidence {
-        *by_pair
-            .entry((item.source.clone(), item.target.clone()))
-            .or_default() += 1;
-    }
-
-    let edges = by_pair
-        .into_iter()
-        .map(|((source, target), count)| Edge {
-            source,
-            target,
-            kind: "runtime".to_string(),
-            evidence: format!("runtime edge evidence count: {count}"),
-        })
-        .collect();
-
-    RuntimeDependencyGraphProjection {
-        projection_rule: RUNTIME_PROJECTION_RULE_VERSION.to_string(),
-        edge_kind: "runtime".to_string(),
-        edges,
-    }
-}
-
-impl PolicyFile {
-    fn read(path: &Path) -> Result<Self, Box<dyn Error>> {
-        let source = fs::read_to_string(path)?;
-        let policy: PolicyFile = serde_json::from_str(&source)?;
-        if policy.schema_version != "signature-policy-v0" {
-            return Err(format!(
-                "unsupported policy schemaVersion: {}",
-                policy.schema_version
-            )
-            .into());
-        }
-        if policy.component_id_kind != COMPONENT_KIND {
-            return Err(format!(
-                "unsupported policy componentIdKind: {}",
-                policy.component_id_kind
-            )
-            .into());
-        }
-        if policy.selector_semantics != "exact-or-prefix-star" {
-            return Err(format!(
-                "unsupported policy selectorSemantics: {}",
-                policy.selector_semantics
-            )
-            .into());
-        }
-        Ok(policy)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct BoundaryPolicy {
-    groups: Vec<BoundaryGroup>,
-    allowed_dependencies: Vec<AllowedDependency>,
-    unmatched_component: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-struct BoundaryGroup {
-    id: String,
-    components: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AllowedDependency {
-    source_group: String,
-    target_group: String,
-    #[allow(dead_code)]
-    reason: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AbstractionPolicy {
-    relations: Vec<AbstractionRelation>,
-    #[allow(dead_code)]
-    unmatched_component: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AbstractionRelation {
-    id: String,
-    abstraction: String,
-    clients: Vec<String>,
-    implementations: Vec<String>,
-    #[serde(default)]
-    allowed_direct_implementation_dependencies: Vec<String>,
-}
-
-fn measure_boundary(
-    boundary: &BoundaryPolicy,
-    components: &[Component],
-    edges: &[Edge],
-) -> Result<(usize, Vec<PolicyViolation>), String> {
-    if boundary.unmatched_component.as_deref() != Some("not-measured") {
-        return Err("boundary.unmatchedComponent must be not-measured".to_string());
-    }
-
-    let component_ids = local_component_ids(components);
-    let mut membership: BTreeMap<String, BTreeSet<String>> = component_ids
-        .iter()
-        .map(|id| (id.clone(), BTreeSet::new()))
-        .collect();
-
-    for group in &boundary.groups {
-        if group.id.is_empty() {
-            return Err("boundary group id must not be empty".to_string());
-        }
-        if group.components.is_empty() {
-            return Err(format!(
-                "boundary group {} has no component selectors",
-                group.id
-            ));
-        }
-
-        for selector in &group.components {
-            let matches = resolve_selector(selector, &component_ids)?;
-            if matches.is_empty() {
-                return Err(format!(
-                    "boundary selector did not match component: {selector}"
-                ));
-            }
-            for component in matches {
-                membership
-                    .get_mut(&component)
-                    .expect("component id came from local set")
-                    .insert(group.id.clone());
-            }
-        }
-    }
-
-    let mut group_by_component = BTreeMap::new();
-    for (component, groups) in membership {
-        if groups.len() != 1 {
-            return Err(format!(
-                "boundary group membership is not unique for {component}"
-            ));
-        }
-        group_by_component.insert(
-            component,
-            groups
-                .into_iter()
-                .next()
-                .expect("exactly one boundary group"),
-        );
-    }
-
-    let allowed: BTreeSet<(String, String)> = boundary
-        .allowed_dependencies
-        .iter()
-        .map(|dependency| {
-            (
-                dependency.source_group.clone(),
-                dependency.target_group.clone(),
-            )
-        })
-        .collect();
-
-    let mut violations = Vec::new();
-    for edge in edges {
-        if !component_ids.contains(&edge.source) || !component_ids.contains(&edge.target) {
-            continue;
-        }
-
-        let source_group = group_by_component
-            .get(&edge.source)
-            .expect("local component has boundary group");
-        let target_group = group_by_component
-            .get(&edge.target)
-            .expect("local component has boundary group");
-        if allowed.contains(&(source_group.clone(), target_group.clone())) {
-            continue;
-        }
-
-        violations.push(PolicyViolation {
-            axis: "boundaryViolationCount".to_string(),
-            source: edge.source.clone(),
-            target: edge.target.clone(),
-            evidence: edge.evidence.clone(),
-            source_group: Some(source_group.clone()),
-            target_group: Some(target_group.clone()),
-            relation_ids: None,
-        });
-    }
-
-    Ok((violations.len(), violations))
-}
-
-fn measure_abstraction(
-    abstraction: &AbstractionPolicy,
-    components: &[Component],
-    edges: &[Edge],
-) -> Result<(usize, Vec<PolicyViolation>), String> {
-    if abstraction.relations.is_empty() {
-        return Err("abstraction.relations must not be empty".to_string());
-    }
-
-    let component_ids = local_component_ids(components);
-    let mut relation_components = Vec::new();
-    for relation in &abstraction.relations {
-        if relation.id.is_empty() {
-            return Err("abstraction relation id must not be empty".to_string());
-        }
-        let clients = resolve_nonempty_selectors(
-            &relation.clients,
-            &component_ids,
-            &format!("abstraction relation {} clients", relation.id),
-        )?;
-        let abstractions = resolve_selector(&relation.abstraction, &component_ids)?;
-        if abstractions.is_empty() {
-            return Err(format!(
-                "abstraction relation {} abstraction selector did not match component: {}",
-                relation.id, relation.abstraction
-            ));
-        }
-        let implementations = resolve_nonempty_selectors(
-            &relation.implementations,
-            &component_ids,
-            &format!("abstraction relation {} implementations", relation.id),
-        )?;
-        let allowed_direct = resolve_optional_selectors(
-            &relation.allowed_direct_implementation_dependencies,
-            &component_ids,
-        )?;
-
-        relation_components.push(ResolvedAbstractionRelation {
-            id: relation.id.clone(),
-            clients,
-            implementations,
-            allowed_direct,
-        });
-    }
-
-    let mut violations = Vec::new();
-    for edge in edges {
-        if !component_ids.contains(&edge.source) || !component_ids.contains(&edge.target) {
-            continue;
-        }
-
-        let relation_ids: Vec<String> = relation_components
-            .iter()
-            .filter(|relation| {
-                relation.clients.contains(&edge.source)
-                    && relation.implementations.contains(&edge.target)
-                    && !relation.allowed_direct.contains(&edge.target)
-            })
-            .map(|relation| relation.id.clone())
-            .collect();
-
-        if relation_ids.is_empty() {
-            continue;
-        }
-
-        violations.push(PolicyViolation {
-            axis: "abstractionViolationCount".to_string(),
-            source: edge.source.clone(),
-            target: edge.target.clone(),
-            evidence: edge.evidence.clone(),
-            source_group: None,
-            target_group: None,
-            relation_ids: Some(relation_ids),
-        });
-    }
-
-    Ok((violations.len(), violations))
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ResolvedAbstractionRelation {
-    id: String,
-    clients: BTreeSet<String>,
-    implementations: BTreeSet<String>,
-    allowed_direct: BTreeSet<String>,
-}
-
-fn local_component_ids(components: &[Component]) -> BTreeSet<String> {
-    components
-        .iter()
-        .map(|component| component.id.clone())
-        .collect()
-}
-
-fn resolve_nonempty_selectors(
-    selectors: &[String],
-    component_ids: &BTreeSet<String>,
-    label: &str,
-) -> Result<BTreeSet<String>, String> {
-    if selectors.is_empty() {
-        return Err(format!("{label} selectors must not be empty"));
-    }
-    let resolved = resolve_optional_selectors(selectors, component_ids)?;
-    if resolved.is_empty() {
-        return Err(format!("{label} selectors did not match any component"));
-    }
-    Ok(resolved)
-}
-
-fn resolve_optional_selectors(
-    selectors: &[String],
-    component_ids: &BTreeSet<String>,
-) -> Result<BTreeSet<String>, String> {
-    let mut resolved = BTreeSet::new();
-    for selector in selectors {
-        let matches = resolve_selector(selector, component_ids)?;
-        if matches.is_empty() {
-            return Err(format!("selector did not match component: {selector}"));
-        }
-        resolved.extend(matches);
-    }
-    Ok(resolved)
-}
-
-fn resolve_selector(
-    selector: &str,
-    component_ids: &BTreeSet<String>,
-) -> Result<BTreeSet<String>, String> {
-    validate_selector(selector)?;
-    if let Some(prefix) = selector.strip_suffix('*') {
-        return Ok(component_ids
-            .iter()
-            .filter(|component| component.starts_with(prefix))
-            .cloned()
-            .collect());
-    }
-
-    Ok(component_ids
-        .iter()
-        .filter(|component| component.as_str() == selector)
-        .cloned()
-        .collect())
-}
-
-fn validate_selector(selector: &str) -> Result<(), String> {
-    if selector.is_empty() {
-        return Err("selector must not be empty".to_string());
-    }
-
-    let star_count = selector.chars().filter(|ch| *ch == '*').count();
-    if star_count > 1 || (star_count == 1 && !selector.ends_with('*')) {
-        return Err(format!("unsupported selector: {selector}"));
-    }
-
-    Ok(())
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ParsedImport {
-    module: String,
-    evidence: String,
-}
-
-fn parse_imports(source: &str) -> Vec<ParsedImport> {
-    let mut imports = Vec::new();
-
-    for line in source.lines() {
-        let trimmed = line.trim();
-        if trimmed.is_empty() || trimmed.starts_with("--") {
-            continue;
-        }
-
-        let Some(rest) = trimmed.strip_prefix("import") else {
-            break;
-        };
-        if !rest.starts_with(char::is_whitespace) {
-            break;
-        }
-
-        let evidence = trimmed
-            .split_once("--")
-            .map(|(before_comment, _)| before_comment.trim())
-            .unwrap_or(trimmed)
-            .to_string();
-        let import_body = evidence
-            .strip_prefix("import")
-            .map(str::trim)
-            .unwrap_or_default();
-
-        for module in import_body.split_whitespace() {
-            imports.push(ParsedImport {
-                module: module.to_string(),
-                evidence: evidence.clone(),
-            });
-        }
-    }
-
-    imports
-}
-
-fn lean_files(root: &Path) -> Result<Vec<PathBuf>, Box<dyn Error>> {
-    let mut files = Vec::new();
-    for entry in WalkDir::new(root)
-        .sort_by_file_name()
-        .into_iter()
-        .filter_entry(|entry| !is_skipped_entry(root, entry))
-    {
-        let entry = entry?;
-        if entry.file_type().is_file() && entry.path().extension() == Some(OsStr::new("lean")) {
-            files.push(entry.into_path());
-        }
-    }
-    Ok(files)
-}
-
-fn is_skipped_entry(root: &Path, entry: &DirEntry) -> bool {
-    let name = entry.file_name().to_string_lossy();
-    if entry.file_type().is_file() {
-        return name == "lakefile.lean";
-    }
-
-    if !entry.file_type().is_dir() {
-        return false;
-    }
-
-    if matches!(name.as_ref(), ".git" | ".lake" | ".elan" | "target") {
-        return true;
-    }
-
-    entry.path().parent() == Some(root) && name == "tools"
-}
-
-fn normalize_relative_path(root: &Path, path: &Path) -> Result<String, Box<dyn Error>> {
-    let relative = path.strip_prefix(root)?;
-    let mut parts = Vec::new();
-    for component in relative.components() {
-        match component {
-            PathComponent::Normal(part) => parts.push(part.to_string_lossy().into_owned()),
-            _ => return Err(format!("unsupported path component in {}", path.display()).into()),
-        }
-    }
-    Ok(parts.join("/"))
-}
-
-fn module_id_from_relative_path(relative_path: &str) -> Result<String, Box<dyn Error>> {
-    let Some(without_ext) = relative_path.strip_suffix(".lean") else {
-        return Err(format!("Lean path does not end with .lean: {relative_path}").into());
-    };
-    Ok(without_ext.replace('/', "."))
-}
-
-fn dedup_edges(import_edges: Vec<ImportEdge>) -> Vec<Edge> {
-    let mut by_pair: BTreeMap<(String, String), String> = BTreeMap::new();
-    for edge in import_edges {
-        by_pair
-            .entry((edge.source, edge.target))
-            .or_insert(edge.evidence);
-    }
-
-    by_pair
-        .into_iter()
-        .map(|((source, target), evidence)| Edge {
-            source,
-            target,
-            kind: "import".to_string(),
-            evidence,
-        })
-        .collect()
-}
-
-fn compute_signature(components: &[Component], edges: &[Edge]) -> Signature {
-    let graph = Graph::from_components_and_edges(components, edges);
-    let sccs = graph.strongly_connected_components();
-    let has_cycle = sccs.iter().any(|scc| {
-        scc.len() > 1
-            || scc
-                .first()
-                .is_some_and(|node| graph.neighbors(node).iter().any(|target| target == node))
-    });
-
-    Signature {
-        has_cycle: usize::from(has_cycle),
-        scc_max_size: sccs.iter().map(Vec::len).max().unwrap_or(0),
-        max_depth: graph.max_bounded_depth(),
-        fanout_risk: graph.unique_edge_count(),
-        boundary_violation_count: 0,
-        abstraction_violation_count: 0,
-    }
-}
-
-#[derive(Debug, Clone)]
-struct Graph {
-    nodes: BTreeSet<String>,
-    adjacency: BTreeMap<String, BTreeSet<String>>,
-}
-
-impl Graph {
-    fn from_components_and_edges(components: &[Component], edges: &[Edge]) -> Self {
-        let mut nodes = BTreeSet::new();
-        let mut adjacency: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
-
-        for component in components {
-            nodes.insert(component.id.clone());
-        }
-
-        for edge in edges {
-            nodes.insert(edge.source.clone());
-            nodes.insert(edge.target.clone());
-            adjacency
-                .entry(edge.source.clone())
-                .or_default()
-                .insert(edge.target.clone());
-        }
-
-        for node in &nodes {
-            adjacency.entry(node.clone()).or_default();
-        }
-
-        Self { nodes, adjacency }
-    }
-
-    fn neighbors(&self, node: &str) -> Vec<String> {
-        self.adjacency
-            .get(node)
-            .map(|neighbors| neighbors.iter().cloned().collect())
-            .unwrap_or_default()
-    }
-
-    fn unique_edge_count(&self) -> usize {
-        self.adjacency.values().map(BTreeSet::len).sum()
-    }
-
-    fn max_fanout(&self) -> usize {
-        self.adjacency
-            .values()
-            .map(BTreeSet::len)
-            .max()
-            .unwrap_or(0)
-    }
-
-    fn reachable_cone_size(&self) -> usize {
-        self.nodes
-            .iter()
-            .map(|node| self.reachable_from(node).len())
-            .max()
-            .unwrap_or(0)
-    }
-
-    fn reachable_from(&self, source: &str) -> BTreeSet<String> {
-        let mut visited = BTreeSet::new();
-        let mut stack = self.neighbors(source);
-        while let Some(node) = stack.pop() {
-            if node == source || !visited.insert(node.clone()) {
-                continue;
-            }
-            stack.extend(self.neighbors(&node));
-        }
-        visited
-    }
-
-    fn strongly_connected_components(&self) -> Vec<Vec<String>> {
-        let mut visited = HashSet::new();
-        let mut order = Vec::new();
-        for node in &self.nodes {
-            self.dfs_order(node, &mut visited, &mut order);
-        }
-
-        let reversed = self.reversed();
-        let mut visited = HashSet::new();
-        let mut sccs = Vec::new();
-        for node in order.iter().rev() {
-            if visited.contains(node) {
-                continue;
-            }
-            let mut scc = Vec::new();
-            reversed.dfs_collect(node, &mut visited, &mut scc);
-            scc.sort();
-            sccs.push(scc);
-        }
-        sccs
-    }
-
-    fn dfs_order(&self, node: &str, visited: &mut HashSet<String>, order: &mut Vec<String>) {
-        if !visited.insert(node.to_string()) {
-            return;
-        }
-
-        for neighbor in self.neighbors(node) {
-            self.dfs_order(&neighbor, visited, order);
-        }
-        order.push(node.to_string());
-    }
-
-    fn dfs_collect(&self, node: &str, visited: &mut HashSet<String>, scc: &mut Vec<String>) {
-        if !visited.insert(node.to_string()) {
-            return;
-        }
-
-        scc.push(node.to_string());
-        for neighbor in self.neighbors(node) {
-            self.dfs_collect(&neighbor, visited, scc);
-        }
-    }
-
-    fn reversed(&self) -> Self {
-        let mut adjacency: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
-        for node in &self.nodes {
-            adjacency.entry(node.clone()).or_default();
-        }
-
-        for (source, targets) in &self.adjacency {
-            for target in targets {
-                adjacency
-                    .entry(target.clone())
-                    .or_default()
-                    .insert(source.clone());
-            }
-        }
-
-        Self {
-            nodes: self.nodes.clone(),
-            adjacency,
-        }
-    }
-
-    fn max_bounded_depth(&self) -> usize {
-        let fuel = self.nodes.len();
-        let mut memo = HashMap::new();
-        self.nodes
-            .iter()
-            .map(|node| self.bounded_depth_from(node, fuel, &mut memo))
-            .max()
-            .unwrap_or(0)
-    }
-
-    fn bounded_depth_from(
-        &self,
-        node: &str,
-        fuel: usize,
-        memo: &mut HashMap<(String, usize), usize>,
-    ) -> usize {
-        if fuel == 0 {
-            return 0;
-        }
-
-        let key = (node.to_string(), fuel);
-        if let Some(depth) = memo.get(&key) {
-            return *depth;
-        }
-
-        let depth = self
-            .neighbors(node)
-            .iter()
-            .map(|target| self.bounded_depth_from(target, fuel - 1, memo) + 1)
-            .max()
-            .unwrap_or(0);
-        memo.insert(key, depth);
-        depth
     }
 }
